+++ contracts/VoucherHub.sol
@@ -238,6 +238,11 @@ contract VoucherHub is
             sponsoredAmount += workerpoolPrice;
         }
         sponsoredAmount = Math.min(balanceOf(msg.sender), sponsoredAmount * volume);
+        // Decrease sponsored amount to make sponsored & non-sponsored amounts
+        // divisible by volume in order to refund plain amounts to voucher and
+        // requester (i.e. make sure there is no hidden remainders) in the event
+        // that tasks are claimed later.
+        sponsoredAmount -= sponsoredAmount % volume;
         if (sponsoredAmount > 0) {
             _burn(msg.sender, sponsoredAmount);
             emit VoucherDebited(msg.sender, sponsoredAmount);